+++ packages/BoSelectDate/index.js
@@ -1,0 +1,7 @@
+import BoSelectDate from './src/main';
+
+BoSelectDate.install = function(Vue) {
+  Vue.component(BoSelectDate.name, BoSelectDate)
+};
+
+export default BoSelectDate;