+++ packages/config/picker-options.js
@@ -1,41 +1,6 @@
-export const pickerOptionsData = (formItems) => {
-    let parseTime = function (time, cFormat) {
-        if (arguments.length === 0) {
-            return null;
-        }
-        const format = cFormat || "{y}-{m}-{d} {h}:{i}:{s}";
-        let date;
-        if (typeof time === "object") {
-            date = time;
-        } else {
-            if (typeof time === "string" && /^[0-9]+$/.test(time)) {
-                time = parseInt(time);
-            }
-            if (typeof time === "number" && time.toString().length === 10) {
-                time = time * 1000;
-            }
-            date = new Date(time);
-        }
-        const formatObj = {
-            y: date.getFullYear(),
-            m: date.getMonth() + 1,
-            d: date.getDate(),
-            h: date.getHours(),
-            i: date.getMinutes(),
-            s: date.getSeconds(),
-            a: date.getDay(),
-        };
-        const time_str = format.replace(/{([ymdhisa])+}/g, (result, key) => {
-            const value = formatObj[key];
-            // Note: getDay() returns 0 on Sunday
-            if (key === "a") {
-                return ["日", "一", "二", "三", "四", "五", "六"][value];
-            }
-            return value.toString().padStart(2, "0");
-        });
-        return time_str;
-    };
+import { parseTime } from '../utils/index'
 
+export const pickerOptionsData = (formItems) => {
     let params = {};
     let pickerOptions = {
         shortcuts: [{
@@ -134,7 +99,16 @@ export const pickerOptionsData = (formItems) => {
             }
 
             pickerOptions = { ...v.pickerOptions, ...pickerOptions }
-        } else if (v.itemType === "multSelect") {
+        } else if (v.itemType === "selectDate") {
+            params[v.prop] = v.value; // 设置默认值
+            v.periodName = v.periodName || (v.prop + '_period');
+            // 判断名称是否已经被使用
+            if(formItems.some(i => i.prop === v.periodName)) {
+                throw('该名称：' + v.periodName + ' 已被使用, 请检查配置');
+            }else {
+                params[v.periodName] = v.periodValue;
+            }
+        }else if (v.itemType === "multSelect") {
             params[v.prop] = v.value || ["_all_"];
         } else {
             params[v.prop] = v.value || "";